move : test for dijkstra and findbridges in algorimsTest
